ANDROID: KVM: arm64: Treat SMCCC function IDs as 32-bit
Commit 128144da229a ("ANDROID: KVM: arm64: Add smc64 trap handling for
protected guests") added SMC handling for selected guests, taking care
to reject calls for function ID falling within the SMCCC "standard
service" range.

However, the check on the function ID is not quite right, as it requires
the upper 32 bits of the register to be zero, whereas the function ID is
32-bit and the upper bits are immaterial.

Check only the bottom 32 bits of the function ID when determining
whether or not it resides in the standard range.

Cc: Nikita Ioffe <ioffe@google.com>
Bug: 556294659

+++ arch/arm64/kvm/hyp/nvhe/pkvm.c
@@ -1971,7 +1971,7 @@ static bool module_handle_guest_trng(struct kvm_vcpu *vcpu)
 }
 
 
-static bool is_standard_secure_service_call(u64 func_id)
+static bool is_standard_secure_service_call(u32 func_id)
 {
 	return (func_id >= PSCI_0_2_FN_BASE && func_id <= ARM_CCA_FUNC_END) ||
 	       (func_id >= PSCI_0_2_FN64_BASE && func_id <= ARM_CCA_64BIT_FUNC_END);
